commiting tests and checkings

+++ CanteenSystem/Application/Logic/ItemLogic.cs
@@ -1,4 +1,5 @@
-﻿using Application.DaoInterfaces;
+﻿using System.Diagnostics.Tracing;
+using Application.DaoInterfaces;
 using Application.LogicInterfaces;
 using Shared.Dtos;
 using Shared.Dtos.IngredientDto;
@@ -13,6 +14,7 @@ public class ItemLogic : IItemLogic
     // private  int value=0;
     private readonly IIngredientDao ingredientDao;
 
+
     public ItemLogic(IItemDao itemDao, IIngredientDao ingredientDao)
     {
         this.itemDao = itemDao;
@@ -21,40 +23,52 @@ public class ItemLogic : IItemLogic
 
 
     public async Task<Item> CreateAsync(ItemCreationDto dto)
-    { 
+    {
         foreach (int ingredientId in dto.IngredientIds)
         {
             Ingredient? existing = await ingredientDao.GetByIdAsync(ingredientId);
-            if (existing == null)
+            if (existing is { Name: null })
             {
                 throw new Exception($"This ingredient you try to use, does not exist!");
             }
 
-            if (existing.Amount == 0)
+            if (existing is { Amount: 0 })
             {
                 throw new Exception($"There is not enough of this ingredient to make this item");
             }
         }
-        
+
+        Item? existingItem = await itemDao.GetByNameAsync(dto.Name);
+        if (existingItem?.Name != null)
+        {
+            throw new Exception("An item with this name already exists");
+        }
+
+        if (dto.IngredientIds == null || !dto.IngredientIds.Any())
+        {
+            throw new Exception("An item needs to have ingredients");
+        }
+
         Item itemCreated = await itemDao.CreateAsync(dto);
+
         return itemCreated;
     }
 
 
-    public Task<IEnumerable<Item>> GetAsync(SearchItemSto searchParameters)
-    {
-        return itemDao.GetAsync(searchParameters);
-    }
-
-    
-
     public async Task DeleteAsync(int id)
     {
         Item? item = await itemDao.GetByIdAsync(id);
-        if (item == null)
+        if (item is {Id:0})
+        {
+            throw new Exception($"Item with ID {id} not found!");
+
+        }
+        if (item== null)
         {
             throw new Exception($"Item with ID {id} was not found!");
         }
+
+        
 
         await itemDao.DeleteAsync(id);
     }
@@ -67,7 +81,7 @@ public class ItemLogic : IItemLogic
             throw new Exception($"Item with ID {id} was not found!");
         }
 
-        return new ItemBasicDto(item.Name, item.Price,item.Ingredients);
+        return new ItemBasicDto(item.Name, item.Price, item.Ingredients);
     }
 
     public async Task<ItemBasicDto?> GetByNameAsync(string name)
@@ -78,7 +92,7 @@ public class ItemLogic : IItemLogic
             throw new Exception($"Item with Name {name} was not found!");
         }
 
-        return new ItemBasicDto(item.Name, item.Price,item.Ingredients);
+        return new ItemBasicDto(item.Name, item.Price, item.Ingredients);
     }
 
     public Task<IEnumerable<Item>> GetAllItemsAsync(SearchItemSto searchParameters)
